huge commit
redefinied Line - due to too much obscursion going on.
removed all submodules, hope they won't crash whole build

+++ Engine/Core/core/OpenGLArrays/ArrayHandler.cpp
@@ -74,7 +74,7 @@ namespace GPU
     }
     bool ArrayHandler::areFreeBuffers()
     {
-        for (int i = 0; i < this->GPUSpace.size(); i++)
+        for (size_t i = 0; i < this->GPUSpace.size(); i++)
         {
             if (!this->GPUSpace[i].Taken())
                 return true;
@@ -83,7 +83,7 @@ namespace GPU
     }
     GPU_Ref *ArrayHandler::getFirstAvailavleBuffer()
     {
-        for (int i = 0; i < GPUSpace.size(); i++)
+        for (size_t i = 0; i < GPUSpace.size(); i++)
         {
             if (!GPUSpace[i].Taken())
                 return GPUSpace[i].Take();
